Split intrinsic gate tests
